pyproject.toml
[project]
name = "notetaker-api"
version = "0.1.0"
description = "A simple note-taker REST API built with FastAPI"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "fastapi[standard]>=0.135.2",
    "pydantic-settings>=2.13.1",
    "sqlalchemy[asyncio]>=2.0.48",
    "uvicorn[standard]>=0.42.0",
]

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "ruff>=0.15.8",
]

# ── FastAPI ──────────────────────────────────────────────────────────────────────
[tool.fastapi]
entrypoint = "app.main:app"

# ── Ruff ──────────────────────────────────────────────────────────────────────
[tool.ruff]
target-version = "py312"
line-length = 88

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "F",   # pyflakes
    "W",   # pycodestyle warnings
    "I",   # isort
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "SIM", # flake8-simplify
]

[tool.ruff.format]
quote-style = "double"

# ── mypy ──────────────────────────────────────────────────────────────────────
[tool.mypy]
python_version = "3.12"
strict = true
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = ["asyncpg.*", "uvicorn.*"]
ignore_missing_imports = true
